requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
flask-restx==1.1.0
mysql-connector-python==8.2.0
PyJWT==2.8.0
pytest==8.0.2
pytest-flask==1.3.0
loguru==0.7.2
Werkzeug==3.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.25
bcrypt==4.0.1
requests==2.31.0 